Exclude new logging API for Net FW, use the previous one.
