+++ dotnet/src/dotnetframework/GxClasses/GxClasses.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
 		<TargetFramework>net462</TargetFramework>
 		<RootNamespace>GxClasses</RootNamespace>
@@ -8,13 +8,8 @@
 		<PackageTags>Data Access</PackageTags>
 		<PackageId>GeneXus.Classes</PackageId>
 	</PropertyGroup>
-	<ItemGroup>
-	  <Compile Include="..\..\dotnetcore\GxClasses\Services\LogService\GXLogService.cs" Link="Services\GXLogService.cs" />
-	</ItemGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Microsoft.Extensions.Logging" Version="7.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="7.0.0" />
 		<PackageReference Include="GeneXus.NodaTime" Version="2.4.19.1" />
 		<PackageReference Include="ManagedFusion.Rewriter" Version="3.7.0" />
 		<PackageReference Include="Microsoft.AspNet.WebPages" Version="3.2.9" />
